backend/requirements.txt
# byeGPT Studio — Backend dependencies
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
pydantic>=2.7.0

# Core byeGPT deps (same as the CLI)
typer>=0.12.0
rich>=13.7.0

# NotebookLM integration
playwright>=1.44.0
notebooklm-py>=0.2.0
